add levels

+++ build.sh
@@ -24,34 +24,42 @@ echo $USERHOME
 
 # TO DO: filter out if CONFIG and SYSCONFIG have same binarys
 # sysconfig - system utils for building
-SYSCONFIG="chown|chmod|adduser|addgroup|egrep|xargs|rm"
+SYSCONFIG="chown|chmod|adduser|addgroup|egrep|xargs|rm|ls"
 # level utils + build utils
 JOINCONFIG="${CONFIG}|${SYSCONFIG}"
 #export CONFIG="$CONFIG"
 #echo "${CONFIG}"
 
-# copy and extract busybox layer
-cp busybox/layer.tar.gz ./ && mkdir layer && tar zxvf layer.tar.gz -C layer 1>/dev/null && rm layer.tar.gz
-
-# rm all unnescesary stuff from busybox
-cd layer/bin && ls | egrep -vw "(${JOINCONFIG})" | xargs rm && ls
-
-# archive new busybox layer and move it to Dockerfile
-cd ../ && pwd && tar zcvf layer.tar.gz ./* && mv layer.tar.gz ../
-
-# return upper and cleanup busybox layer
-cd ../ && rm -rf ./layer
-
-# archive game layer and move it to dockerfile
-cd level${level}/layer && tar zcvf level.tar.gz ./* && mv level.tar.gz ../../ && cd ../../
-
-# build and push image
-docker build -t ctf.school:5000/suzenescape/${NAME} . \
+if [ -e level${level}/Dockerfile ]; then
+  cd level${level}/
+  docker build -t ctf.school:5000/suzenescape/${NAME} . \
   --build-arg USERNAME=${NAME} \
   --build-arg CONFIG=${CONFIG} \
   --build-arg USERHOME=${USERHOME}
-#  --build-arg JOINCONFIG=${CONFIG}
-docker push ctf.school:5000/suzenescape/${NAME}
+  docker push ctf.school:5000/suzenescape/${NAME}
+else
+  # copy and extract busybox layer
+  cp busybox/layer.tar.gz ./ && mkdir layer && tar zxvf layer.tar.gz -C layer 1>/dev/null && rm layer.tar.gz
 
-# cleanup
-rm layer.tar.gz level.tar.gz
+  # rm all unnescesary stuff from busybox
+  cd layer/bin && ls | egrep -vw "(${JOINCONFIG})" | xargs rm && ls
+
+  # archive new busybox layer and move it to Dockerfile
+  cd ../ && pwd && tar zcvf layer.tar.gz ./* && mv layer.tar.gz ../
+
+  # return upper and cleanup busybox layer
+  cd ../ && rm -rf ./layer
+
+  # archive game layer and move it to dockerfile
+  cd level${level}/layer && tar zcvf level.tar.gz ./* && mv level.tar.gz ../../ && cd ../../
+
+  # build and push image
+  docker build -t ctf.school:5000/suzenescape/${NAME} . \
+  --build-arg USERNAME=${NAME} \
+  --build-arg CONFIG=${CONFIG} \
+  --build-arg USERHOME=${USERHOME}
+  docker push ctf.school:5000/suzenescape/${NAME}
+
+  # cleanup
+  rm layer.tar.gz level.tar.gz
+fi
